requirements.txt
gensim==4.3.1
mysql_connector_repackaged==0.3.1
pandas==2.0.1
Requests==2.31.0
streamlit==1.22.0
